Add attribution and address pr reviews
