+++ build.gradle.kts
@@ -96,16 +96,6 @@ allprojects {
     }
 }
 
-val androidLibs = setOf(
-    "sentry-android-core",
-    "sentry-android-ndk",
-    "sentry-android-fragment",
-    "sentry-android-navigation",
-    "sentry-android-okhttp",
-    "sentry-android-timber",
-    "sentry-compose-android"
-)
-
 subprojects {
     plugins.withId(Config.QualityPlugins.detektPlugin) {
         configure<DetektExtension> {
@@ -156,26 +146,7 @@ subprojects {
 
             @Suppress("UnstableApiUsage")
             configure<MavenPublishBaseExtension> {
-                pom {
-                    withXml {
-                        // workaround for https://github.com/gradle/gradle/issues/3170
-                        val dependencies = asNode().children().find {
-                            it is Node && it.name().toString().endsWith("dependencies")
-                        } as Node?
-
-                        dependencies?.children()?.forEach { dep ->
-                            if (dep !is Node) {
-                                return@forEach
-                            }
-                            val artifactId = dep.children().first {
-                                it is Node && it.name().toString().endsWith("artifactId")
-                            } as Node
-                            if ((artifactId.children()[0] as String) in androidLibs) {
-                                dep.appendNode("type", "aar")
-                            }
-                        }
-                    }
-                }
+                assignAarTypes()
             }
 
             // maven central info go to:
